agregado md5 a pass de encargados y sepracion de menus dependiendo del usuario
