+++ main.py
@@ -1,0 +1,6 @@
+import os
+from menu import menu
+os.system('cls')
+
+
+menu()
