+++ test.py
@@ -1,11 +1,13 @@
-#cifrar en sha256
+#cifrar en md5
 import hashlib
 
 def cifrar(password):
-    hashlib_obj = hashlib.sha256()
-    hashlib.update(password.encode('utf-8'))
+    hashlib_obj = hashlib.md5()
+    hashlib_obj.update(password.encode('utf-8'))
     cifrado = hashlib_obj.hexdigest()
-    return cifrado.hexdigest()
+    #print(cifrado, len(cifrado))
+    return cifrado
+
 
 
 #menu()
